Update templates to include public topics
- All users of the site, no matter if they aren't registered, can view
the public topics and entries of other users.

- The buttons for adding and editing entries of a topic only appear if
the currently logged in user is the owner of the topic, otherwise the
entries can just be viewed.

- When the user is loggen in, the color of their public entries appear
yellow and there is a new message warning that the changes the make to
any public entry can be viewed by other users of the site.